feat: swa insert implemented in mail report
