+++ controllers/home.ts
@@ -1,5 +1,4 @@
 import { Request, Response } from "express";
-import { mailReportsConsumer } from "../consumers/mail-reports-consumer";
 import logger from '../logger/logger';
 
 // GET '/'
@@ -11,9 +10,8 @@ const healthcheck = async (_req: Request, res: Response) => {
 const test = async (req: Request, res: Response) => {
     try {
         const result: any = req.body.data;
-        const output = await mailReportsConsumer(result);
         
-        return res.send(output);
+        return res.send(result);
     } catch (err) {
         logger.error(err);
         return res.status(500).send(err);
