+++ package.json
@@ -43,7 +43,6 @@
     "express": "^4.18.1",
     "firebase-admin": "^11.0.0",
     "google-libphonenumber": "^3.2.30",
-    "google-protobuf": "^3.21.2",
     "helmet": "^5.1.0",
     "jshashes": "^1.0.8",
     "jsonwebtoken": "^8.5.1",
